test(canon): keep handler fixtures under length gate
